追記

+++ select.php
@@ -1,11 +1,13 @@
 <!-- php開始 -->
 <?php
+session_start();
 //funcs.php読み込み
 include("funcs.php");
+sschk();
 
 //DB接続
 try {
-  $pdo = new PDO('mysql:dbname=kadai09;charset=utf8;host=localhost','root','');
+  $pdo = new PDO('mysql:dbname=coralbadger93_kadai09;charset=utf8;host=mysql57.coralbadger93.sakura.ne.jp','coralbadger93','suwayuki10');
 } catch (PDOException $e) {
   exit('DBConnectError:'.$e->getMessage());
 }
@@ -39,7 +41,13 @@ $values =  $stmt->fetchAll(PDO::FETCH_ASSOC);
 <body>
 <!-- header作成 -->
 <header>
-  <p class="select_p">データ一覧表示</p>
+  <p class="select_p">データ一覧</p>
+  <div class="select_p1">[<?=$_SESSION["name"]?>さん、こんにちは]</div>
+  <div class="select_a1">
+    <a class="select_a2" href="menu.php">メニュー</a>
+    <a class="select_a2" href="index.php">データ登録画面</a>
+    <a class="select_a2" href="logout.php">ログアウト</a>
+  </div>
 </header>
 <!-- テーブル作成 -->
 <div>
@@ -50,7 +58,9 @@ $values =  $stmt->fetchAll(PDO::FETCH_ASSOC);
       <th>URL</th>
       <th>コメント</th>
       <th>デート</th>
+      <?php if($_SESSION["kanri_flg"]==1){?>
       <th>削除</th>
+      <?php }?>
     </tr>
       <?php foreach($values as $v){ ?>
     <tr>
@@ -59,15 +69,13 @@ $values =  $stmt->fetchAll(PDO::FETCH_ASSOC);
       <td class="select_td1"><?=$v["url"]?></a></td>
       <td class="select_td2"><?=$v["comment"]?></a></td>
       <td class="select_td1"><?=$v["indate"]?></td>
-      <td class="select_td1"><a class="select_id" href="delete.php?id=<?=$v["id"]?>">🗑️</a></td>
+      <?php if($_SESSION["kanri_flg"]==1){?>
+        <td class="select_td1"><a class="select_id" href="delete.php?id=<?=$v["id"]?>">🗑️</a></td>
+      <?php }?>
     </tr>
       <?php }?>
   </table>
 </div>
 <!-- footer作成 -->
-<!-- index.phpへ移動 -->
-  <div class="select_a1">
-    <a class="select_a2" href="index.php">データ登録</a>
-  </div>
 </body>
 </html>
